+++ learn-go-with-pocket-sized-projects/03-log-story/go.mod
@@ -1,3 +1,3 @@
-module alexioannides/go-playpen/learn-go-with-pocket-sized-projects/03-log-story/logger
+module alexioannides/go-playpen/learn-go-with-pocket-sized-projects/03-log-story
 
 go 1.22.5
