+++ learn-go-with-pocket-sized-projects/03-log-story/pocketlog/level.go
